requirements.txt
faiss_cpu==1.9.0
numpy==1.24.3
pandas==2.2.3
sentence_transformers==3.2.1
streamlit==1.39.0
~treamlit==1.15.2
